portal/brownie_api/src/requirements.txt
flask
requests
Flask-Cors
ipfshttpclient==0.8.0a2
eth-brownie
web3
